requirements/local.txt
# Local development dependencies go here
-r base.txt

# Django environ
django-environ==0.4.5
django-debug-toolbar==2.2
mixer